pyproject.toml
[tool.poetry]
name = "nacometa-backend"
version = "0.1.0"
description = ""
authors = ["Nikita Derugin <dafun@icloud.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
aiohttp = "3.7.2"
aiohttp_cors = "0.7.0"
aioredis = "1.3.1"
alembic = "1.3.0"
psycopg2-binary = "^2.9.3"
SQLAlchemy = "1.3.19"
graphene = "2.1.8"
aiohttp-graphql = "1.0.0"
graphql-server-core = "1.1.1"
graphql-core = "2.3.2"
graphql-ws = "0.3.0"
trafaret-config = "2.0.2"
trafaret = "2.0.0"
gino = "1.0.1"
Authlib = "0.13"
argon2-cffi = "19.2.0"
PyJWT = "1.7.1"
boto3 = "1.24.55"
APScheduler = "3.6.3"
zenpy = "2.0.20"
openpyxl = "3.0.5"
loguru = "0.5.3"
aiohttp-pydantic = "1.9.1"
facebook-business = "11.0.0"
Pillow = "8.3.1"
frozendict = "2.0.6"
phonenumbers = "8.12.41"
fastapi = "0.67.0"
uvicorn = {version = "0.14.0", extras = ["standard"]}
fastapi-utils = "0.2.1"
python-multipart = "0.0.5"
fastapi-pagination = "0.8.2"
shortuuid = "1.0.1"
paypal-checkout-serversdk = "1.0.1"
paypalrestsdk = "1.13.1"
httpx = "0.23.3"
NudeNet = "2.0.9"
nest-asyncio = "1.5.4"
virtualenv = "20.10.0"
gunicorn = "20.1.0"
SQLAlchemy-Utils = "0.38.2"
sqlalchemy-stubs = "^0.4"
types-six = "^1.16.15"
types-requests = "^2.27.25"
types-PyYAML = "^6.0.7"
types-pytz = "^2021.3.7"
types-frozendict = "^2.0.8"
types-python-dateutil = "^2.8.14"
furl = "^2.1.3"
google-cloud-translate = "^3.8.4"
localization = { git = "https://git.webpp.ru/nacometa1/localization-python-backend.git", rev = "0.4.8"}
dependency-injector = "^4.41.0"
gino-starlette = "0.1.2"

sentry-sdk = {extras = ["fastapi"], version = "^1.7.2"}
pip = "^22.2.2"
pybind11 = "^2.10.0"
fasttext-wheel = "^0.9.2"
multimethod = "^1.9"
numpy = "1.23.5"

[tool.poetry.dev-dependencies]
pytest = "7.2.0"
mypy = "^0.942"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = ">=3.8"
coverage = "^6.3.2"
flake8-logging-format = "^0.7.5"
flake8-alfred = "^1.1.1"
flake8-bugbear = "^22.3.23"
flake8-builtins = "^1.5.3"
flake8-coding = "^1.3.2"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.8.0"
flake8-debugger = "^4.0.0"
flake8-eradicate = "^1.2.0"
flake8-multiline-containers = "^0.0.18"
flake8-pep3101 = "^1.3.0"
flake8-print = "^4.0.0"
flake8-pytest-style = "^1.6.0"
flake8-return = "^1.1.3"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.2"
flake8-SQL = "0.3.0"
pep8-naming = ">0.1"
flake8-cognitive-complexity = "^0.1.0"
flake8-isort = "^4.1.1"
flake8-annotations = "^2.8.0"
types-requests = "^2.27.16"
flake8-docstrings = "^1.6.0"
pytest-asyncio = "0.20.1"
pytest-env = "0.8.1"
factory-boy = "3.2.1"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
env = ["TESTING=True", "TZ=UTC", "GS_ENVIRONMENT=test"]
filterwarnings = [
    "ignore:.*.Arguments instead of.*:DeprecationWarning",
    "ignore:.*'collections' instead of from 'collections.abc'.*:DeprecationWarning",
    "ignore:.*'context' alias has been deprecated.*:DeprecationWarning",
    "ignore:.*'variables' alias has been deprecated.*:DeprecationWarning",
    "ignore:.*\"algorithms\" argument when calling decode().*:DeprecationWarning",
    "ignore:.*deprecated alias for the builtin `float`.*:DeprecationWarning",
]

[tool.black]
line-length = 79
include = '\.py$'
exclude = '''
(
    /(
        \.git
      | \.mypy_cache
      | env
      | test-reports
      | ptvs_virtualenv_proxy.py
      | migrations
      | venv
    )/
    | goldstream/__init__.py
)

'''

[tool.isort]
line_length = 79
py_version=39
skip="migrations"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
profile="black"
extend_skip = ["./tests/rest/test_dreams.py", "venv"]
